Add Spring Boot 2 support
Both Spring Boot 1.5 and Spring Boot 2 are supported.

Improvements on modules and structure: `togglz-spring-boot-starter`
module now combines:

* `togglz-spring-common-configurer`: support for common components
* `togglz-spring-boot1-configurer`: support for c 1.5 specific components
* `togglz-spring-boot2-configurer`: support for Spring Boot 2.x specific components

Spring boot specific modules are conditionally loaded by Spring Boot using
@ConditionalOnClass. For Spring Boot 1.x, we check for AbstractEndpoint
class and for Spring Boot 2, we check for Endpoint annotation class.

Those classes only exist in their versions, so they're ok to make
sure that we load the right AutoConfiguration classes.

Also, packages were renamed to ensure that they are not shared by
different modules (for Java 9 compatibility).

Fixes: #271
