--- spring-boot-starter/pom.xml
@@ -1,139 +1,0 @@
-<?xml version="1.0"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-
-  <parent>
-    <groupId>org.togglz</groupId>
-    <artifactId>togglz-project</artifactId>
-    <version>2.6.0-SNAPSHOT</version>
-  </parent>
-
-  <artifactId>togglz-spring-boot-starter</artifactId>
-  <name>Togglz - Spring Boot Starter</name>
-  <description>Togglz - Spring Boot Starter</description>
-
-  <properties>
-    <spring-boot.version>1.5.4.RELEASE</spring-boot.version>
-    <thymeleaf-togglz.version>1.0.1.RELEASE</thymeleaf-togglz.version>
-  </properties>
-
-  <dependencyManagement>
-     <dependencies>
-        <dependency>
-            <!-- Import dependency management from Spring Boot -->
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-dependencies</artifactId>
-            <version>${spring-boot.version}</version>
-            <type>pom</type>
-            <scope>import</scope>
-        </dependency>
-    </dependencies>
-  </dependencyManagement>
-
-  <developers>
-    <developer>
-      <id>marceloverdijk</id>
-      <email>maceloverdijk@gmail.com</email>
-      <name>Marcel Overdijk</name>
-      <timezone>CET</timezone>
-    </developer>
-  </developers>
-
-  <dependencies>
-
-    <dependency>
-      <groupId>org.togglz</groupId>
-      <artifactId>togglz-core</artifactId>
-    </dependency>
-
-    <dependency>
-      <groupId>org.togglz</groupId>
-      <artifactId>togglz-console</artifactId>
-      <optional>true</optional>
-    </dependency>
-
-    <dependency>
-      <groupId>org.togglz</groupId>
-      <artifactId>togglz-spring-core</artifactId>
-    </dependency>
-
-    <dependency>
-      <groupId>org.togglz</groupId>
-      <artifactId>togglz-spring-web</artifactId>
-    </dependency>
-
-    <dependency>
-      <groupId>org.togglz</groupId>
-      <artifactId>togglz-spring-security</artifactId>
-      <optional>true</optional>
-    </dependency>
-
-    <dependency>
-      <groupId>org.springframework.boot</groupId>
-      <artifactId>spring-boot-actuator</artifactId>
-      <optional>true</optional>
-    </dependency>
-
-    <dependency>
-        <groupId>org.springframework.boot</groupId>
-        <artifactId>spring-boot-configuration-processor</artifactId>
-        <optional>true</optional>
-    </dependency>
-
-    <dependency>
-      <groupId>org.springframework.security</groupId>
-      <artifactId>spring-security-config</artifactId>
-      <optional>true</optional>
-    </dependency>
-
-    <dependency>
-      <groupId>org.springframework.security</groupId>
-      <artifactId>spring-security-web</artifactId>
-      <optional>true</optional>
-    </dependency>
-
-   <dependency>
-      <groupId>org.springframework</groupId>
-      <artifactId>spring-webmvc</artifactId>
-      <optional>true</optional>
-    </dependency>
-
-    <dependency>
-      <groupId>com.github.heneke.thymeleaf</groupId>
-      <artifactId>thymeleaf-extras-togglz</artifactId>
-      <version>${thymeleaf-togglz.version}</version>
-      <optional>true</optional>
-    </dependency>
-
-    <dependency>
-      <groupId>org.hibernate</groupId>
-      <artifactId>hibernate-validator</artifactId>
-      <optional>true</optional>
-    </dependency>
-
-    <dependency>
-      <groupId>org.jboss.spec.javax.servlet</groupId>
-      <artifactId>jboss-servlet-api_3.0_spec</artifactId>
-      <scope>provided</scope>
-    </dependency>
-
-    <dependency>
-      <groupId>junit</groupId>
-      <artifactId>junit</artifactId>
-      <optional>true</optional>
-    </dependency>
-
-    <dependency>
-      <groupId>org.springframework</groupId>
-      <artifactId>spring-test</artifactId>
-      <optional>true</optional>
-    </dependency>
-
-    <dependency>
-        <artifactId>spring-boot-starter-test</artifactId>
-        <groupId>org.springframework.boot</groupId>
-        <scope>test</scope>
-    </dependency>
-
-  </dependencies>
-</project>
