--- spring-boot-starter/src/main/resources/META-INF/spring.factories
@@ -1,10 +1,0 @@
-# Auto Configure
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-org.togglz.spring.boot.autoconfigure.TogglzAutoConfiguration
-
-org.springframework.boot.actuate.autoconfigure.ManagementContextConfiguration=\
-org.togglz.spring.boot.autoconfigure.TogglzManagementContextConfiguration
-
-# Test Execution Listeners
-org.springframework.test.context.TestExecutionListener=\
-org.togglz.spring.boot.test.TogglzTestExecutionListener
