--- spring-boot-starter/template.mf
@@ -1,16 +1,0 @@
-Manifest-Version: 1.0
-Bundle-ManifestVersion: 2
-Bundle-SymbolicName: org.togglz.spring
-Bundle-Vendor: http://www.togglz.org/
-Bundle-Version: ${osgi.bundles.version}
-Bundle-Name: ${project.name}
-Version-Patterns:
- default;pattern="[=.=.=, =.+1.0)"
-Import-Template:
- com.github.heneke.thymeleaf.togglz.*;version="0",
- javax.servlet.*;version="0",
- javax.validation.*;version="0",
- org.springframework.*;version="[3.0,4.0)",
- org.springframework.boot.*;version="0",
- org.springframework.security.*;version="[2.0,4.0)",
- org.togglz.*;version="${osgi.bundles.version:default}"
