+++ spring-boot/common-autoconfigure/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,7 @@
+# Auto Configure
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+org.togglz.spring.boot.autoconfigure.TogglzAutoConfiguration
+
+# Test Execution Listeners
+org.springframework.test.context.TestExecutionListener=\
+org.togglz.spring.boot.test.TogglzTestExecutionListener
