+++ spring-boot/common-autoconfigure/src/test/resources/features-file/features.properties
@@ -1,0 +1,2 @@
+FEATURE_ONE = true
+FEATURE_TWO = false
